+++ src/boards/bonadrone.hpp
@@ -48,9 +48,9 @@ namespace hf {
             static const LSM6DSM::Rate_t   AODR   = LSM6DSM::ODR_1660Hz;
             static const LSM6DSM::Rate_t   GODR   = LSM6DSM::ODR_1660Hz;
 
-            // Biases computed by Simon using Juan & Pep's LSM6DSM/Examples/Calibrate sketch
+            // Gyro bias will be estimated by the ESKF filter
             float ACCEL_BIAS[3] = {-0.020306,0.008926,0.029526};
-            float GYRO_BIAS[3]  = {0.301350,-0.818594,-0.701652};
+            float GYRO_BIAS[3]  = {0.0,0.0,0.0};
 
 
             // Instance variables -----------------------------------------------------------------------------------
@@ -141,8 +141,6 @@ namespace hf {
 
                 delay(100);
 
-                // Calibrate IMU on startup
-                _lsm6dsm.calibrate(GYRO_BIAS, ACCEL_BIAS);
                 // Clear the interrupt
                 _lsm6dsm.clearInterrupt();
                 setLed(false);