+++ src/filters/eskf.hpp
@@ -105,7 +105,13 @@ namespace hf {
           dptr += nn;
           eskfp.tmp7 = dptr;
           dptr += nn;
-          eskfp.tmp8 = dptr;        
+          eskfp.tmp8 = dptr;
+          dptr += m;
+          eskfp.tmp9 = dptr;
+          dptr += ne*ne;
+          eskfp.tmp10 = dptr;
+          dptr += ne*ne;
+          eskfp.tmp11 = dptr;
       }
       
       void synchState(void)
@@ -142,18 +148,58 @@ namespace hf {
           eskfp.x[1] = 0.0;
           eskfp.x[2] = 0.0;
           eskfp.x[3] = 0.0;
+          eskfp.x[4] = 0.301350;
+          eskfp.x[5] = -0.818594;
+          eskfp.x[6] = -0.701652;
           
-          eskfp.P[0] = 1.0;
-          eskfp.P[1] = 0;
-          eskfp.P[2] = 0;
+          // First Column
+          eskfp.P[0]  =  1.0;
+          eskfp.P[6]  =  0.0;
+          eskfp.P[12] = 0.0;
+          eskfp.P[18] = 0.0;
+          eskfp.P[24] = 0.0;
+          eskfp.P[30] = 0.0;
           
-          eskfp.P[3] = 0.0;
-          eskfp.P[4] = 1.0;
-          eskfp.P[5] = 0.0;
-        
-          eskfp.P[6] = 0.0;
-          eskfp.P[7] = 0.0;
-          eskfp.P[8] = 1.0;
+          // Second Column
+          eskfp.P[1]  = 0.0;
+          eskfp.P[7]  =  1.0;
+          eskfp.P[13] = 0.0;
+          eskfp.P[19] = 0.0;
+          eskfp.P[25] = 0.0;
+          eskfp.P[31] = 0.0;
+
+          // Third Column
+          eskfp.P[2]  = 0.0;
+          eskfp.P[8]  = 0.0;
+          eskfp.P[14] = 1.0;
+          eskfp.P[20] = 0.0;
+          eskfp.P[26] = 0.0;
+          eskfp.P[32] = 0.0;
+
+          // Fourth Column
+          eskfp.P[3]  = 0.0;
+          eskfp.P[9]  = 0.0;
+          eskfp.P[15] = 0.0;
+          eskfp.P[21] = 1.0;
+          eskfp.P[27] = 0.0;
+          eskfp.P[33] = 0.0;
+          
+          // Fifth Column
+          eskfp.P[4]  = 0.0;
+          eskfp.P[10] = 0.0;
+          eskfp.P[16] = 0.0;
+          eskfp.P[22] = 0.0;
+          eskfp.P[28] = 1.0;
+          eskfp.P[34] = 0.0;
+          
+          // Sixth Column
+          eskfp.P[5]  = 0.0;
+          eskfp.P[11] = 0.0;
+          eskfp.P[17] = 0.0;
+          eskfp.P[23] = 0.0;
+          eskfp.P[29] = 0.0;
+          eskfp.P[35] = 1.0;
+
       }
 
       void addSensorESKF(ESKF_Sensor * sensor)
@@ -183,7 +229,7 @@ namespace hf {
           dt = (t_now - t_lastCall)/1000000.0f;
           t_lastCall = t_now;
           
-          sensor->getJacobianModel(eskfp.Fx, dt);
+          sensor->getJacobianModel(eskfp.Fx, eskfp.x, dt);
           sensor->getJacobianErrors(eskfp.Fdx, dt);
           sensor->getCovarianceEstimation(eskfp.Q);
           
@@ -247,31 +293,60 @@ namespace hf {
           mulvec(eskfp.K, eskfp.hx, eskfp.dx, errorStates, observations);
           
           /* P_k = P_k - K_k Z_k K^T_k  */
-          transpose(eskfp.K, eskfp.Kt, observations, errorStates);
-          mulmat(eskfp.K, eskfp.tmp3, eskfp.tmp0, errorStates, observations, errorStates);
-          mulmat(eskfp.tmp0, eskfp.Kt, eskfp.tmp3, errorStates, errorStates, observations);
-          sub(eskfp.Pp, eskfp.tmp3, eskfp.tmp0, errorStates);
-          makesym(eskfp.tmp0, eskfp.P, errorStates);
+          //transpose(eskfp.K, eskfp.Kt, observations, errorStates);
+          //mulmat(eskfp.K, eskfp.tmp3, eskfp.tmp0, errorStates, observations, errorStates);
+          //mulmat(eskfp.tmp0, eskfp.Kt, eskfp.tmp3, errorStates, errorStates, observations);
+          //sub(eskfp.Pp, eskfp.tmp3, eskfp.tmp0, errorStates);
+          //makesym(eskfp.tmp0, eskfp.P, errorStates);
+          
+          /* P = (I-KH)*P*(I-KH)' + KZK' */
+          mulmat(eskfp.K, eskfp.H, eskfp.tmp9, errorStates, observations, errorStates); // K*H
+          negate(eskfp.tmp9, errorStates, errorStates); // -K*H
+          mat_addeye(eskfp.tmp9, errorStates); // -K*H + I
+          transpose(eskfp.tmp9, eskfp.tmp10, errorStates, errorStates); // (-K*H + I)'
+          mulmat(eskfp.tmp9, eskfp.Pp, eskfp.tmp11, errorStates, errorStates, errorStates); // (-K*H + I)*P
+          mulmat(eskfp.tmp11, eskfp.tmp10, eskfp.P, errorStates, errorStates, errorStates); // (-K*H + I)*P*(-K*H + I)'
+          // Z is stored in eskf.tmp3 and K in eskfp.K
+          transpose(eskfp.K, eskfp.Kt, errorStates, observations); // K'
+          mulmat(eskfp.tmp3, eskfp.Kt, eskfp.tmp2, observations, observations, errorStates); // Z*K'
+          mulmat(eskfp.K, eskfp.tmp2, eskfp.tmp0, errorStates, observations, errorStates); // K*Z*K'
+          accum(eskfp.P, eskfp.tmp0, errorStates, errorStates); 
+
+          // Serial.println("P before G update");
+          // printMatrix(eskfp.Pp, errorStates, errorStates);
           
           /* Error injection */
-          eskfp.tmp6[0] = 1.0;
-          eskfp.tmp6[1] = eskfp.dx[0]/2.0;
-          eskfp.tmp6[2] = eskfp.dx[1]/2.0;
-          eskfp.tmp6[3] = eskfp.dx[2]/2.0;
+          // XXX Quaternion injection as a method
+          float tmp[4];
+          tmp[0] = 1.0;
+          tmp[1] = eskfp.dx[0]/2.0;
+          tmp[2] = eskfp.dx[1]/2.0;
+          tmp[3] = eskfp.dx[2]/2.0;
           Quaternion::computeqL(eskfp.qL, eskfp.fx);
-          mulvec(eskfp.qL, eskfp.tmp6, eskfp.tmp7, nominalStates, nominalStates);
-          norvec(eskfp.tmp7, eskfp.x, nominalStates);
+          mulvec(eskfp.qL, tmp, eskfp.tmp7, 4, 4);
+          norvec(eskfp.tmp7, eskf.x, 4); // XXX Be careful when the quat is not the first state
+
+          eskfp.x[4] += eskfp.dx[3];
+          eskfp.x[5] += eskfp.dx[4];
+          eskfp.x[6] += eskfp.dx[5];
 
           /* Update covariance*/
-          eskfp.tmp5[0] = eskfp.dx[0]/2.0;
+          /*eskfp.tmp5[0] = eskfp.dx[0]/2.0;
           eskfp.tmp5[1] = eskfp.dx[1]/2.0;
           eskfp.tmp5[2] = eskfp.dx[2]/2.0;
-          skew(eskfp.tmp5, eskfp.G);
+          newSkew(eskfp.tmp5, eskfp.G);
           negate(eskfp.G, errorStates, errorStates);
           mat_addeye(eskfp.G, errorStates);
           transpose(eskfp.G, eskfp.tmp0, errorStates, errorStates);
-          mulmat(eskfp.P, eskfp.tmp0, eskfp.Pp, errorStates, errorStates, errorStates);
-          mulmat(eskfp.G, eskfp.Pp, eskfp.P, errorStates, errorStates, errorStates);
+          mulmat(eskfp.Pp, eskfp.tmp0, eskfp.tmp9, errorStates, errorStates, errorStates);
+          mulmat(eskfp.G, eskfp.tmp9, eskfp.tmp10, errorStates, errorStates, errorStates);
+          
+          // Force its symmetry: P = (P + P')/2
+          makesym(eskfp.tmp10, eskfp.P, errorStates);*/
+          
+          
+          // Serial.println("P after G update");
+          // printMatrix(eskfp.P, errorStates, errorStates);
           
           /* reset error state */
           zeros(eskfp.dx, errorStates, 1);
@@ -281,7 +356,7 @@ namespace hf {
       }
 
       // XXX Debug
-      void printMatrix3(float M[3][3], int r, int c)
+      void printMatrix(float * M, int r, int c)
       {
           for (int ii=0; ii<r; ++ii)
           {
@@ -289,31 +364,11 @@ namespace hf {
             {
               if (jj == c-1)
               {
-                Serial.println(M[ii][jj],8);
+                Serial.println(M[ii*c+jj],8);
               }
               else
               {
-                Serial.print(M[ii][jj],8);
-                Serial.print(",");
-              }
-              
-            }
-          }
-      }
-      
-      void printMatrix4(float M[4][4], int r, int c)
-      {
-          for (int ii=0; ii<r; ++ii)
-          {
-            for (int jj=0; jj<c; ++jj)
-            {
-              if (jj == c-1)
-              {
-                Serial.println(M[ii][jj],8);
-              }
-              else
-              {
-                Serial.print(M[ii][jj],8);
+                Serial.print(M[ii*c+jj],8);
                 Serial.print(",");
               }